+++ js/main.js
@@ -1,2 +1,7 @@
-import './util.js';
-import './render.js';
+import {generateAds} from './render.js';
+import {offersArray} from './offers-array.js';
+
+const similarOffer = document.querySelector('#map-canvas');
+
+generateAds(offersArray);
+similarOffer.appendChild(generateAds(offersArray));
