+++ js/util.js
@@ -1,5 +1,3 @@
-import {FEATURES, PHOTOS, HOUSING_TYPES, CHECKIN_ARRAY, CHECKOUT_ARRAY, LATITUDE, LONGITUDE, PRICE, ROOMS, GUESTS} from './data.js';
-
 const getRandomInteger = (min, max) => {
 
   if (min < 0 || max < 0) {
@@ -33,44 +31,9 @@ const getShuffledArray = (array) => {
   return array;
 };
 
-const getLocation =  () =>  ({
-  lat: getRandomFloatInteger(LATITUDE.min, LATITUDE.max, 5),
-  lng: getRandomFloatInteger(LONGITUDE.min, LONGITUDE.max, 5),
+const getLocation =  (lat, lng) =>  ({
+  lat: getRandomFloatInteger(lat.min, lat.max, 5),
+  lng: getRandomFloatInteger(lng.min, lng.max, 5),
 });
 
-const getOffersArray = (amount) => {
-  const offerArray = [];
-  for (let i = 1; i <= amount; i++) {
-    const currentLocation = getLocation();
-    offerArray.push({
-      author: {
-        avatar: `img/avatars/user0${getRandomInteger(1, 8)}.png`,
-      },
-
-      offer: {
-        title: 'Аренда жилья в Токио',
-        address: `${currentLocation.lat}, ${currentLocation.lng}`,
-        price: getRandomInteger(PRICE.min, PRICE.max),
-        type: getRandomArrayElement(getShuffledArray(HOUSING_TYPES)),
-        rooms: getRandomInteger(ROOMS.min, ROOMS.max),
-        guests: getRandomInteger(GUESTS.min, GUESTS.max),
-        checkin: getRandomArrayElement(CHECKIN_ARRAY),
-        checkout: getRandomArrayElement(CHECKOUT_ARRAY),
-        features: getShuffledArray(FEATURES).slice(0, getRandomInteger(1, FEATURES.length)),
-        description: 'Чистое, светлое помещение для комфортного проживания с отличной инфраструктурой.',
-        photos: getShuffledArray(PHOTOS).slice(0, getRandomInteger(1, PHOTOS.length)),
-      },
-
-      location: {
-        lat: currentLocation.lat,
-        lng: currentLocation.lng,
-      },
-    });
-  }
-  return offerArray;
-};
-
-const offersArray = getOffersArray(1);
-offersArray;
-
-export {offersArray};
+export {getRandomInteger, getRandomFloatInteger, getRandomArrayElement, getShuffledArray, getLocation};
